Everything You Need to Know About CIPA Compliance
Congress enacted the Children's Internet Protection Act (CIPA) in 2000 to address issues relating to minors' access to harmful or obscene content via the Internet. CIPA compliance is mandatory to libraries and schools that leverage the E-Rate program to get discounts for Internet connections or access. These organizations are mandated to provide a safe environment for kids when they go online.


https://www.nuedusec.com/cipa-compliance.php